refact/front/LoginWGoogleBtn.jsx: moves to src/components/buttons/LoginWGoogleBtn.jsx
