+++ frontend/src/components/buttons/LoginWGoogleBtn.jsx
@@ -1,6 +1,6 @@
 import { PropTypes } from 'prop-types';
 import googleSVG from '../assets/google.svg';
-import RoundedBtn from './buttons/RoundedBtn';
+import RoundedBtn from './RoundedBtn';
 
 const LoginWGoogleBtn = ({text}) => {
   const handleLogin = () => {
