+++ frontend/src/pages/Login.jsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import  authenticateUser  from '../store/requests/authenticateUser';
-import LoginWGoogleBtn from '../components/LoginWGoogleBtn';
+import LoginWGoogleBtn from '../components/buttons/LoginWGoogleBtn';
 import ButtonGreen from '../components/buttons/ButtonGreen';
 import Input from '../components/Input';
